Extract Hanami::CLI::ShellHelper (#execute)
From Hanami::CLI::FileHelper, since `execute` never belonged there
