+++ spec/unit/hanami/cli/file_helper_spec.rb
@@ -213,17 +213,4 @@ RSpec.describe Hanami::CLI::FileHelper do
       end
     end
   end
-
-  describe "#execute" do
-    it "executes command" do
-      expect(Kernel).to receive(:system).with(
-        "arbitrary_command",
-        Hash[argument: 1]
-      )
-      expect(stdout).to receive(:puts).with(
-        "         run  arbitrary_command\n"
-      )
-      subject.execute("arbitrary_command", argument: 1)
-    end
-  end
 end
